TIP-874: move job related classes
