+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/job_constraints.yml
@@ -1,7 +1,7 @@
 parameters:
-    pim_catalog.job.job_parameters.constraint_collection_provider.compute_product_models_descendants.class: Pim\Component\Catalog\Job\JobParameters\ConstraintCollectionProvider\ComputeProductModelsDescendants
-    pim_catalog.job.job_parameters.constraint_collection_provider.compute_completeness_of_products_family.class: Pim\Component\Catalog\Job\JobParameters\ConstraintCollectionProvider\ComputeCompletenessOfProductsFamily
-    pim_catalog.job.job_parameters.constraint_collection_provider.compute_family_variant_structure_changes.class: Pim\Component\Catalog\Job\JobParameters\ConstraintCollectionProvider\ComputeFamilyVariantStructureChanges
+    pim_catalog.job.job_parameters.constraint_collection_provider.compute_product_models_descendants.class: Akeneo\Pim\Enrichment\Component\Product\Job\JobParameters\ConstraintCollectionProvider\ComputeProductModelsDescendants
+    pim_catalog.job.job_parameters.constraint_collection_provider.compute_completeness_of_products_family.class: Akeneo\Pim\Enrichment\Component\Product\Job\JobParameters\ConstraintCollectionProvider\ComputeCompletenessOfProductsFamily
+    pim_catalog.job.job_parameters.constraint_collection_provider.compute_family_variant_structure_changes.class: Akeneo\Pim\Enrichment\Component\Product\Job\JobParameters\ConstraintCollectionProvider\ComputeFamilyVariantStructureChanges
 
 services:
     pim_catalog.job.job_parameters.constraint_collection_provider.compute_product_models_descendants:
