+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/jobs.yml
@@ -1,7 +1,7 @@
 parameters:
-    pim_catalog.tasklet.compute_product_models_descendants.class: Pim\Component\Catalog\Job\ComputeProductModelsDescendantsTasklet
-    pim_catalog.tasklet.compute_completeness_of_products_family.class: Pim\Component\Catalog\Job\ComputeCompletenessOfProductsFamilyTasklet
-    pim_catalog.tasklet.compute_family_variant_structure_changes.class: Pim\Component\Catalog\Job\ComputeFamilyVariantStructureChangesTasklet
+    pim_catalog.tasklet.compute_product_models_descendants.class: Akeneo\Pim\Enrichment\Component\Product\Job\ComputeProductModelsDescendantsTasklet
+    pim_catalog.tasklet.compute_completeness_of_products_family.class: Akeneo\Pim\Enrichment\Component\Product\Job\ComputeCompletenessOfProductsFamilyTasklet
+    pim_catalog.tasklet.compute_family_variant_structure_changes.class: Akeneo\Pim\Enrichment\Component\Product\Job\ComputeFamilyVariantStructureChangesTasklet
 
     pim_catalog.compute_product_models_descendants.job_type: 'compute_product_models_descendants'
     pim_catalog.compute_product_models_descendants.job_name: 'compute_product_models_descendants'
